@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	  <TargetFrameworks>netstandard2.0;net6.0;net7.0;net8.0</TargetFrameworks>
+	  <TargetFrameworks>net7.0;net8.0</TargetFrameworks>
 	  <Description>One more library for MediatR. It allows you to configure the most commonly used PipelineBehaviors in a very simple way.</Description>
 	  <Authors>Pascal LIBAUD</Authors>
 	  <Company></Company>
@@ -20,6 +20,7 @@
     <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.9.0" />
     <PackageReference Include="MediatR" Version="12.2.0" />
     <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="8.0.0" />
+    <PackageReference Include="One.More.Lib.For.Linq" Version="0.0.1" />
     <PackageReference Include="Polly" Version="8.3.0" />
     <None Include="..\..\LICENSE.txt">
       <Pack>True</Pack>
@@ -28,5 +29,4 @@
 
     <None Include="..\..\readme.md" Pack="true" PackagePath="\" />
   </ItemGroup>
-
 </Project>
