@@ -3,30 +3,30 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using FluentValidation;
+using One.More.Lib.For.Linq;
 using MediatR;
 
-namespace One.More.Lib.For.MediatR
+namespace One.More.Lib.For.MediatR;
+
+internal class FluentValidationPipelineBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
 {
-    internal class FluentValidationPipelineBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
+    private readonly IEnumerable<IValidator<TRequest>> _validators;
+
+    public FluentValidationPipelineBehavior(IEnumerable<IValidator<TRequest>> validators)
     {
-        private readonly IEnumerable<IValidator<TRequest>> _validators;
+        _validators = validators;
+    }
 
-        public FluentValidationPipelineBehavior(IEnumerable<IValidator<TRequest>> validators)
-        {
-            _validators = validators;
-        }
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
+    {
+        var failures = await _validators.OmSelectAsync(v => v.ValidateAsync(request, cancellationToken))
+            .OmSelectManyAsync(result => result.Errors)
+            .OmWhereAsync(f => f != null)
+            .OmToListAsync();
 
-        public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
-        {
-            var failures = (await Task.WhenAll(_validators.Select(v => v.ValidateAsync(request, cancellationToken))))
-                .SelectMany(result => result.Errors)
-                .Where(f => f != null)
-                .ToList();
+        if (failures.Any())
+            throw new ValidationException(failures);
 
-            if (failures.Any())
-                throw new ValidationException(failures);
-
-            return await next();
-        }
+        return await next();
     }
 }
