@@ -3,57 +3,56 @@ using FluentValidation;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace One.More.Lib.For.MediatR
+namespace One.More.Lib.For.MediatR;
+
+public static class ServiceCollectionExtensions
 {
-    public static class ServiceCollectionExtensions
+    public static IServiceCollection AddMediatRExtensions(this IServiceCollection services, Action<MediatRExtensionConfiguration> configuration)
     {
-        public static IServiceCollection AddMediatRExtensions(this IServiceCollection services, Action<MediatRExtensionConfiguration> configuration)
+        var serviceConfig = new MediatRExtensionConfiguration();
+
+        configuration(serviceConfig);
+
+        if (serviceConfig.PerformanceLoggerSupport)
         {
-            var serviceConfig = new MediatRExtensionConfiguration();
+            services.AddSingleton(new PerformanceLoggerConfiguration { TriggerThreshold = serviceConfig.TriggerThreshold });
 
-            configuration(serviceConfig);
-
-            if (serviceConfig.PerformanceLoggerSupport)
-            {
-                services.AddSingleton(new PerformanceLoggerConfiguration { TriggerThreshold = serviceConfig.TriggerThreshold });
-
-                services.Add(new ServiceDescriptor(typeof(IPipelineBehavior<,>), typeof(PerformanceLoggerPipelineBehavior<,>), serviceConfig.Lifetime));
-            }
-
-            if (serviceConfig.MemoryCacheSupport)
-            {
-                services.AddMemoryCache();
-
-                services.AddSingleton(new MemoryCacheConfiguration
-                {
-                    AbsoluteExpiration = serviceConfig.AbsoluteExpiration,
-                    AbsoluteExpirationRelativeToNow = serviceConfig.AbsoluteExpirationRelativeToNow,
-                    Priority = serviceConfig.Priority,
-                    SlidingExpiration = serviceConfig.SlidingExpiration,
-                });
-
-                services.Add(new ServiceDescriptor(typeof(IPipelineBehavior<,>), typeof(MemoryCachePipelineBehavior<,>), serviceConfig.Lifetime));
-            }
-
-            if (serviceConfig.FluentValidationSupport)
-            {
-                services.AddValidatorsFromAssemblies(serviceConfig.Assemblies, serviceConfig.Lifetime, serviceConfig.Filter, serviceConfig.IncludeInternalTypes);
-
-                services.Add(new ServiceDescriptor(typeof(IPipelineBehavior<,>), typeof(FluentValidationPipelineBehavior<,>), serviceConfig.Lifetime));
-            }
-
-            if (serviceConfig.RetrySupport)
-            {
-                services.AddSingleton(new RetryConfiguration
-                {
-                    RetryCount = serviceConfig.RetryCount,
-                    RetryDelay = serviceConfig.RetryDelay
-                });
-
-                services.Add(new ServiceDescriptor(typeof(IPipelineBehavior<,>), typeof(RetryPipelineBehavior<,>), serviceConfig.Lifetime));
-            }
-
-            return services;
+            services.Add(new ServiceDescriptor(typeof(IPipelineBehavior<,>), typeof(PerformanceLoggerPipelineBehavior<,>), serviceConfig.Lifetime));
         }
+
+        if (serviceConfig.MemoryCacheSupport)
+        {
+            services.AddMemoryCache();
+
+            services.AddSingleton(new MemoryCacheConfiguration
+            {
+                AbsoluteExpiration = serviceConfig.AbsoluteExpiration,
+                AbsoluteExpirationRelativeToNow = serviceConfig.AbsoluteExpirationRelativeToNow,
+                Priority = serviceConfig.Priority,
+                SlidingExpiration = serviceConfig.SlidingExpiration,
+            });
+
+            services.Add(new ServiceDescriptor(typeof(IPipelineBehavior<,>), typeof(MemoryCachePipelineBehavior<,>), serviceConfig.Lifetime));
+        }
+
+        if (serviceConfig.FluentValidationSupport)
+        {
+            services.AddValidatorsFromAssemblies(serviceConfig.Assemblies, serviceConfig.Lifetime, serviceConfig.Filter, serviceConfig.IncludeInternalTypes);
+
+            services.Add(new ServiceDescriptor(typeof(IPipelineBehavior<,>), typeof(FluentValidationPipelineBehavior<,>), serviceConfig.Lifetime));
+        }
+
+        if (serviceConfig.RetrySupport)
+        {
+            services.AddSingleton(new RetryConfiguration
+            {
+                RetryCount = serviceConfig.RetryCount,
+                RetryDelay = serviceConfig.RetryDelay
+            });
+
+            services.Add(new ServiceDescriptor(typeof(IPipelineBehavior<,>), typeof(RetryPipelineBehavior<,>), serviceConfig.Lifetime));
+        }
+
+        return services;
     }
 }
